Sprint-1 Unit testing shifted in a separate file for Shifat
